feat: uses Unit8Array in place of Buffer
